# docs/swagger.yaml
basePath: /api/
definitions:
...
  AuthorizeRequest:
    properties:
      address:
        type: string
      signature:
        type: string
    required:
      - address
      - signature
    type: object
...
  ChangeUserNameRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  GetMeResponse:
    properties:
      data:
        $ref: '#/definitions/GetUserData'
      status:
        type: string
    type: object
...
  GetUserData:
    properties:
      user:
        $ref: '#/definitions/User'
    type: object
...
  ResponseError:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  ResponseSuccess:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  User:
    properties:
      balance:
        type: number
      created_at:
        type: string
      debt:
        type: number
      email:
        type: string
      exclusive_code:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      last_requested_at:
        type: string
      media_qualities_config:
        type: string
      updated_at:
        type: string
      wallet_address:
        type: string
      wallet_connection:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Aioz Stream Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aioz Stream API
  version: "1.0"
paths:
...
  /user/challenge/{address}:
    get:
      consumes:
        - application/json
        - application/x-www-form-urlencoded
      description: Challenge for wallet verify. This will return a challenge string
        for signing
      parameters:
        - description: Link wallet request
          in: path
          name: address
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: success
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Get user challenge
      tags:
        - user
  /user/link_wallet:
    put:
      consumes:
        - application/json
        - application/x-www-form-urlencoded
      description: Link wallet with existing account by signature and address from
        metamask
      parameters:
        - description: Link wallet request
          in: body
          name: AuthorizeRequest
          required: true
          schema:
            $ref: '#/definitions/AuthorizeRequest'
      produces:
        - application/json
      responses:
        "200":
          description: success
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Link wallet
      tags:
        - user
  /user/me:
    get:
      consumes:
        - application/json
        - application/x-www-form-urlencoded
      description: get current user's information
      parameters:
        - description: authorization
          in: header
          name: Authorization
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/GetMeResponse'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - BasicAuth: [ ]
        - Bearer: [ ]
      summary: Get me
      tags:
        - user
  /user/rename:
    patch:
      consumes:
        - application/json
        - application/x-www-form-urlencoded
      parameters:
        - description: Full name
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/ChangeUserNameRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - BearerAuth: [ ]
      summary: Change full name
      tags:
        - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"